% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KAMA.R
\name{KAMA_localization}
\alias{KAMA_localization}
\title{KAMA Localization Analysis}
\usage{
KAMA_localization(
  kappa_set,
  tau_set,
  M,
  q = 0.1,
  genetic_info = NULL,
  study_label = NULL,
  seed = 111
)
}
\arguments{
\item{kappa_set}{A numeric matrix of kappa statistics. Rows correspond to variables,
and columns correspond to studies.
\strong{Note:} The order of variables (rows) and studies (columns) must strictly match that of \code{tau_set}.}

\item{tau_set}{A numeric matrix of tau statistics. Rows correspond to variables,
and columns correspond to studies.
\strong{Note:} The order of variables (rows) and studies (columns) must strictly match that of \code{kappa_set}.}

\item{M}{Integer. The number of knockoffs used.}

\item{q}{Numeric. The FDR control level (default 0.1).}

\item{genetic_info}{Optional data frame containing variable metadata.}

\item{study_label}{Character vector. Custom labels for the studies (columns). Defaults to "study1", "study2", etc.}

\item{seed}{Integer. Random seed.}
}
\value{
A data frame of the significant variables containing:
\itemize{
\item \code{genetic_info}: (If provided) The appended metadata.
\item \code{kappa_set}: The kappa statistics.
\item \code{tau_set}: The tau statistics.
\item \code{qvalue}: KAMA's q-value for meta-analysis FDR control.
\item \code{localization}: A character column specifying the studies/cohorts where the variable is significant.
}
}
\description{
The main pipeline for KAMA to identify study-specific associations.
It performs meta-analysis and localizes the signal to specific studies/cohorts
}
