% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KAMA.R
\name{KAMA_meta_analysis}
\alias{KAMA_meta_analysis}
\title{KAMA Meta-Analysis}
\usage{
KAMA_meta_analysis(kappa_set, tau_set, M, genetic_info = NULL, seed = 111)
}
\arguments{
\item{kappa_set}{A numeric matrix of kappa statistics. Rows correspond to variables,
and columns correspond to studies.
\strong{Note:} The order of variables (rows) and studies (columns) must strictly match that of \code{tau_set}.}

\item{tau_set}{A numeric matrix of tau statistics. Rows correspond to variables,
and columns correspond to studies.
\strong{Note:} The order of variables (rows) and studies (columns) must strictly match that of \code{kappa_set}.}

\item{M}{Integer. The number of knockoffs used. Must be >= max(kappa_set).}

\item{genetic_info}{Optional data frame or matrix containing additional information (e.g., SNP IDs, positions) to be appended to the results.}

\item{seed}{Integer. Random seed for reproducibility (default 111).}
}
\value{
A data frame containing:
\itemize{
\item \code{genetic_info} (if provided)
\item \code{kappa_set} columns
\item \code{tau_set} columns
\item \code{qvalue}: KAMA's q-value for meta-analysis FDR control.
}
}
\description{
KAMA Meta-Analysis
}
